test(distributed-locks): lift composite TestLease fake into Fakes
Mechanical move of the nested acquire-path lease fake so the reader-writer and
semaphore composite suites can share it instead of copy-pasting. Call sites are
renamed; no assertion changes (79 before, 79 after).
